Refactor invoice processing and add max threads property
Removed the code line that created a new thread for processing invoices in UVNetworkManager.java and instead submitted the task to executor service. Also, a new property was added to mini.properties for setting the maximum number of threads, providing more flexibility and responsiveness to varying system conditions. This property allows the control of maximum possible threads managing network events.