Replace log method with print_log in UVNetworkManager
The print_log method has been added to both classes, UVNode and UVNetworkManager, and is used to print messages in these classes. This commits reduces the interruptions to the execution flow by replacing certain instances of log with print_log. The logging of some messages has been switched to debugging, for instance, the liquidity reservation log in UVNode. Furthermore, a transaction presentation improvement is added to UVTimechain for better readability.