Increase network thread limit and adjust bootstrap parameters
The maximum number of network threads has been increased from 3000 to 4000. The properties for bootstrapping a network including the number of nodes and the block periods have been updated as well. Additionally, bootstrap median and mean time have been reformatted as a fraction of the total bootstrap period for more accurate node startup time calculations.