Implement fatal exit on duplicate channel addition
The code has been updated to execute a fatal exit if there is an attempt to add a duplicate channel within the 'addChannel' and 'addAnnouncedChannel' methods. Previously, the program only displayed a warning message. The modification strengthens data integrity by maintaining a single unique instance of a channel within our channel set.