Add comments in LNInvoice class
Added explanatory comments about the 'min_cltv_expiry' field in LNInvoice class. These comments explain the role and purpose of this field, clarifying that it's set by the final recipient and used to calculate the CLTV expiry for the final hop, to aid better understanding for future code maintenance and enhancements.